added randomPoint function to utility, changed Block constructor to float, float

+++ src/main/java/it/unibo/geosurv/core/Block.java
@@ -6,7 +6,7 @@ import java.awt.Rectangle;
 
 public class Block extends GameObject {
 
-    public Block(int x, int y, ID id) {
+    public Block(float x, float y, ID id) {
         super(x, y, id);
     }
